# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YouTube to MP3 Converter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VideoInfoRequest {
  /** YouTube video URL */
  url: string;
}

export interface VideoInfoResponse {
  title: string;
  thumbnail: string;
  /** Duration in seconds */
  duration: number;
}

/**
 * MP3 bitrate quality
 */
export type ConvertRequestQuality =
  (typeof ConvertRequestQuality)[keyof typeof ConvertRequestQuality];

export const ConvertRequestQuality = {
  NUMBER_128: "128",
  NUMBER_192: "192",
  NUMBER_320: "320",
} as const;

export interface ConvertRequest {
  /** YouTube video URL */
  url: string;
  /** MP3 bitrate quality */
  quality: ConvertRequestQuality;
}

export interface ConvertResponse {
  success: boolean;
  title: string;
  /** Download URL path */
  download: string;
}

export interface ErrorResponse {
  error: string;
}