# lib/api-zod/src/generated/types/convertRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YouTube to MP3 Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { ConvertRequestQuality } from "./convertRequestQuality";

export interface ConvertRequest {
  /** YouTube video URL */
  url: string;
  /** MP3 bitrate quality */
  quality: ConvertRequestQuality;
}