# lib/api-zod/src/generated/types/convertResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YouTube to MP3 Converter API
 * OpenAPI spec version: 0.1.0
 */

export interface ConvertResponse {
  success: boolean;
  title: string;
  /** Download URL path */
  download: string;
}